fix: resolve Kotlin compilation errors blocking CI build
- PredictionEngine.kt: fix type mismatch in `when` expression – UP/DOWN
  branches returned Float while `else` returned Double, causing the
  inferred type to be {Comparable<*> & Number} instead of Double.
  Add explicit .toDouble() on all Float branches to match
  PredictionResult.predictedPrice: Double.

- EmptyState.kt: add missing EmptyState composable in ui/components
  (AlertsScreen.kt referenced it but it was never defined).

- AlertsScreen.kt: add import for the new EmptyState component.

Agent-Logs-Url: https://github.com/upamanyu92/FinanceAI-android/sessions/904cb8fb-8a9c-4ee3-9b9d-9142916466ed

+++ app/src/main/java/com/stocksense/app/engine/PredictionEngine.kt
@@ -126,8 +126,8 @@ class PredictionEngine(private val context: Context) {
         val confidence = probs[directionIdx]
         val lastClose = closes.last()
         val predictedPrice = when (direction) {
-            "UP" -> lastClose * (1 + confidence * 0.02f)
-            "DOWN" -> lastClose * (1 - confidence * 0.02f)
+            "UP" -> (lastClose * (1 + confidence * 0.02f)).toDouble()
+            "DOWN" -> (lastClose * (1 - confidence * 0.02f)).toDouble()
             else -> lastClose.toDouble()
         }
